Move reducer specs into separate files
