+++ src/js/modules/comments/reducer.spec.ts
@@ -1,0 +1,3 @@
+describe('Comments reducer', () => {
+  it('TODO');
+});
