+++ src/js/modules/previews/reducer.spec.ts
@@ -1,0 +1,3 @@
+describe('Previews reducer', () => {
+  it('TODO');
+});
