+++ src/js/modules/streaming/reducer.spec.ts
@@ -1,0 +1,3 @@
+describe('Streaming reducer', () => {
+  it('TODO');
+});
