+++ src/js/modules/user/reducer.spec.ts
@@ -1,0 +1,3 @@
+describe('User reducer', () => {
+  it('TODO');
+});
